Move the test as suggested
